+++ src/main/java/com/space/server/auth/service/implementation/AdditionalInfoUpdater.java
@@ -5,10 +5,10 @@ import com.space.server.auth.presentation.dto.request.AdditionalInfoRequest;
 import com.space.server.common.jwt.exception.InvalidTokenException;
 import com.space.server.common.jwt.exception.RefreshTokenNotFoundException;
 import com.space.server.common.jwt.util.JwtUtil;
-import com.space.server.user.domain.Users;
-import com.space.server.user.domain.repository.UserRepository;
-import com.space.server.user.domain.value.Role;
-import com.space.server.user.exception.UserNotFoundException;
+import com.space.server.domain.user.domain.Users;
+import com.space.server.domain.user.domain.repository.UserRepository;
+import com.space.server.domain.user.domain.value.Role;
+import com.space.server.domain.user.exception.UserNotFoundException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
