+++ src/main/java/com/space/server/common/jwt/util/AuthenticationUtil.java
@@ -2,7 +2,7 @@ package com.space.server.common.jwt.util;
 
 import com.space.server.auth.service.dto.CustomUserDetails;
 import com.space.server.oauth.service.dto.CustomOAuth2User;
-import com.space.server.user.domain.value.Role;
+import com.space.server.domain.user.domain.value.Role;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
