+++ src/main/java/com/space/server/domain/chat/service/CommandChatService.java
@@ -26,8 +26,8 @@ import com.space.server.domain.state.domain.value.Status;
 import com.space.server.domain.state.service.implementation.StateCreator;
 import com.space.server.domain.state.service.implementation.StateReader;
 import com.space.server.domain.state.service.implementation.StateUpdater;
-import com.space.server.user.domain.Users;
-import com.space.server.user.service.implementation.UserReader;
+import com.space.server.domain.user.domain.Users;
+import com.space.server.domain.user.service.implementation.UserReader;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
