+++ src/main/java/com/space/server/domain/chat/service/QueryChatService.java
@@ -5,7 +5,7 @@ import com.space.server.domain.chat.presentation.dto.response.ChatResponse;
 import com.space.server.domain.chat.service.implementation.ChatReader;
 import com.space.server.domain.quiz.service.implementation.QuizReader;
 import com.space.server.domain.state.service.implementation.StateReader;
-import com.space.server.user.service.implementation.UserReader;
+import com.space.server.domain.user.service.implementation.UserReader;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
