+++ src/main/java/com/space/server/domain/inventory/domain/repository/InventoryRepository.java
@@ -3,7 +3,7 @@ package com.space.server.domain.inventory.domain.repository;
 import com.space.server.domain.inventory.domain.Inventory;
 import com.space.server.domain.item.domain.Item;
 import com.space.server.domain.item.domain.value.Category;
-import com.space.server.user.domain.Users;
+import com.space.server.domain.user.domain.Users;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
