+++ src/main/java/com/space/server/domain/inventory/service/CommandInventoryService.java
@@ -5,8 +5,8 @@ import com.space.server.domain.item.domain.Item;
 import com.space.server.domain.item.domain.value.Category;
 import com.space.server.domain.item.service.implementation.ItemReader;
 import com.space.server.domain.inventory.service.implementation.*;
-import com.space.server.user.domain.Users;
-import com.space.server.user.domain.repository.UserRepository;
+import com.space.server.domain.user.domain.Users;
+import com.space.server.domain.user.domain.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
