+++ src/main/java/com/space/server/domain/state/domain/repository/StateRepository.java
@@ -2,7 +2,7 @@ package com.space.server.domain.state.domain.repository;
 
 import com.space.server.domain.quiz.domain.Quiz;
 import com.space.server.domain.state.domain.State;
-import com.space.server.user.domain.Users;
+import com.space.server.domain.user.domain.Users;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
