+++ src/main/java/com/space/server/domain/state/presentation/dto/response/StateResponse.java
@@ -3,7 +3,7 @@ package com.space.server.domain.state.presentation.dto.response;
 import com.space.server.domain.quiz.domain.Quiz;
 import com.space.server.domain.state.domain.value.Status;
 import com.space.server.domain.state.domain.State;
-import com.space.server.user.domain.Users;
+import com.space.server.domain.user.domain.Users;
 
 public record StateResponse(
         Long stateId,
